Wait for RUN bit to be cleared when stopping
When the RUN bit in the PCM_CONTROL register is cleared, we need to read it
back and wait until the engine has really stopped. Otherwise there is a
change for a rogue DMA write to appear after the stop.

Also reset the engine completely when the ALSA device is closed, and create
a shorthand for this operation to make the call-sites more compact.