Add network driver
This patch makes the driver create a TUN-like network interface on
probe. The data this TUN emits are received from a ring buffer that
is shared with the PCI card's userspace, and the packets it received
are pushed to another ringbuffer.

Interrupts are latched on both sides to inform the respective other
side of new data in the ringbuffer.

Ther performance of this device is not overwhelming, but ~25 MBit/s
should suffice for the setup work we are doing.